fix the movement in mini map without checked invalid read and leaks

+++ src/includes/graphics.h
@@ -6,7 +6,7 @@
 /*   By: lslater <lslater@42angouleme.fr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/18 13:07:13 by lslater           #+#    #+#             */
-/*   Updated: 2024/09/19 13:00:05 by lslater          ###   ########.fr       */
+/*   Updated: 2024/09/26 13:33:28 by basverdi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int		init_graphique(t_data *data);
 void	gen_minimap(t_mlx *mlx, t_data *data);
 int		clear_minimap(t_mlx *mlx);
 
-void	move_player(t_mlx *mlx, int y, int x);
+void	move_player(t_mlx *mlx, float y, float x);
 
 int		init_window(t_mlx *mlx, t_data *data);
 int		windowmap_hook(int event, void* param);
